requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openpyxl==3.1.2
Pillow==10.1.0
python-dotenv==1.0.0
reportlab==4.0.7
requests==2.31.0
urllib3==2.1.0
waitress==2.1.2
Werkzeug==3.0.1
